Fix: intro, sign in naming
